Remove update of ozv_statistics on all servers before each run of MonLocalJobs
If there are a lot of servers, the update of every ovz_statistics lasts
very long and so the monitoring kills itself.

If a global update of all ovz_statistics should be done, then it should be
done separately and with checking of the timestamp. It's not the concern
of the MonSystem. If the timestamp is to old (older than a minute) the
MonSystem gets its new ovz_statistics by itself.