Changes in MonLocalBehaviors due to mon_behavior_params
Change MonLocalBehaviorInterface to give additional params to dynamically
access the ovzStatistics array. So the keypath can be configured for each
MonLocalJob in the model/db. So only one Behavior is needed for Physical
and Virtual. Implement those new behaviors.

In this state it is possible to make multiple Diskspacefree MonJobs which
can be configured to watch specific partitions/mountpoints.