Bugfixing to use Disk in GB and Memory in MB (#43)
Jobs ovz_new_vs and ovz_modify_vs have now diskspace in GB and memory in
MB in params. the conversion of bytes, say GB to MB to use in prlctl is
now done exactly where it is needed: when concatenating the command for
the prlctl set.

On the other side the controller and view in the phalcon app does now work
with the official format in the model. Helpers are used for the
conversion. One thing does not work well, it is the form which is
generated directly from the model. in this case the raw data is set as
values in the form and the user sees directly GB or MB without suffix and
maybe to big numbers for usability. For this case a new issue will be
opened.